+++ components/shys_m5_dial/ha_device_mode_light_brightness.h
@@ -6,7 +6,7 @@ namespace esphome
     {
         class HaDeviceModeLightBrightness: public esphome::shys_m5_dial::HaDeviceMode {
             protected:
-                void sendValueToHomeAssistant(float value) override {
+                void sendValueToHomeAssistant(int value) override {
                     haApi.turnLightOn(this->device.getEntityId(), value);
                 }
 
@@ -57,7 +57,7 @@ namespace esphome
                         if(this->isValueModified()){
                             return;
                         }
-                        auto val = parse_number<float>(state);
+                        auto val = parse_number<int>(state);
                         if (!val.has_value()) {
                             this->setReceivedValue(0);
                             ESP_LOGD("HA_API", "No Brightness value in %s for %s", state.c_str(), this->device.getEntityId().c_str());
@@ -68,7 +68,7 @@ namespace esphome
                     });
                 }
 
-                bool onTouch(M5DialDisplay& display, float x, float y) override {
+                bool onTouch(M5DialDisplay& display, uint16_t x, uint16_t y) override {
                     return this->defaultOnTouch(display, x, y);
                 }
